backend/requirements.txt
numpy==2.3.5
fastapi
uvicorn[standard]
python-multipart
supabase
python-dotenv
pydantic
pydantic-settings
email-validator>=2.0.0
dnspython>=2.0.0
langgraph
langchain
langchain-openai
langchain-core
openai
pytz
cryptography
PyJWT
requests
httpx
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-storage==2.10.0
google-generativeai==0.8.5
google-genai==1.39.0
groq==0.4.1
python-dateutil
Pillow
numpy
scikit-learn==1.8.0
aiofiles
razorpay
beautifulsoup4
aiohttp
textstat
apscheduler
playwright>=1.40.0
python-docx
PyPDF2
